API-IRIS-Django2.0/iris/requirements.txt
Django==2.0
joblib==0.17.0
numpy==1.19.2
pandas==1.1.3
python-dateutil==2.8.1
python-decouple==3.3
pytz==2020.1
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sklearn==0.0
threadpoolctl==2.1.0
